Update README.md
This is a mission statement encapsulated in a life plan dovetailed with a spiritual quest to find the fortune that my logic is worth in a world that is fighting to create the most useful applications for human use. I know humans. I am one. I know what they need. What they want and what my work is now. I am enlightened and I hope you find the same enlightenment as me in your path as humans. I am a poet first and a scientist second but first and foremost I am fluent in Midwest English and we do things perfectly. Peep game and watch me rise.